requirements.txt

torch
torchaudio
openai-whisper
pydub
numpy